+++ Sources/VComponents/Components/State Pickers/Toggle Button (Rectangular)/VRectangularToggleButton.swift
@@ -156,11 +156,12 @@ public struct VRectangularToggleButton<Label>: View where Label: View {
         icon: Image
     ) -> some View {
         icon
-            .resizable()
-            .scaledToFit()
+            .applyIf(uiModel.isIconResizable, transform: { $0.resizable() })
+            .applyIfLet(uiModel.iconContentMode, transform: { $0.aspectRatio(nil, contentMode: $1) })
+            .applyIfLet(uiModel.iconColors, transform: { $0.foregroundStyle($1.value(for: internalState)) })
+            .applyIfLet(uiModel.iconOpacities, transform: { $0.opacity($1.value(for: internalState)) })
+            .font(uiModel.iconFont)
             .frame(size: uiModel.iconSize)
-            .foregroundStyle(uiModel.iconColors.value(for: internalState))
-            .opacity(uiModel.iconOpacities.value(for: internalState))
     }
 
     private var backgroundView: some View {
@@ -258,7 +259,7 @@ extension VRectangularToggleButtonInternalState {
                 uiModel: {
                     var uiModel: VRectangularToggleButtonUIModel = .init()
                     uiModel.backgroundColors.off = uiModel.backgroundColors.pressedOff
-                    uiModel.iconColors.off = uiModel.iconColors.pressedOff
+                    uiModel.iconColors!.off = uiModel.iconColors!.pressedOff // Force-unwrap
                     return uiModel
                 }(),
                 state: .constant(.off),
@@ -278,7 +279,7 @@ extension VRectangularToggleButtonInternalState {
                 uiModel: {
                     var uiModel: VRectangularToggleButtonUIModel = .init()
                     uiModel.backgroundColors.on = uiModel.backgroundColors.pressedOn
-                    uiModel.iconColors.on = uiModel.iconColors.pressedOn
+                    uiModel.iconColors!.on = uiModel.iconColors!.pressedOn // Force-unwrap
                     return uiModel
                 }(),
                 state: .constant(.on),